Limit editing of a ticket to a user
