+++ backend/app.js
@@ -11,7 +11,7 @@ const schema = require('./graphql/index.js');
 const authModel = require('./models/auth.js'); // For authentication
 
 const fetchTrainPositions = require('./models/trains.js');
-const editTicket = require('./models/test.js');
+const editTicket = require('./models/edits.js');
 
 const port = process.env.PORT || 1337;
 const app = express();
